# ClientSDKs/StoreApi/openapi.g.yaml
openapi: 3.0.1
paths:
  /StoreModule/pet/listPets:
    get:
      summary: List all pets
      operationId: StoreModuleListPets
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        400:
          description: Invalid status value
      x-lz-gencall: PetRepo.ListAsync(callerInfo)
  /StoreModule/pet/findByStatus:
    get:
      summary: Finds Pets by status
      description: Multiple status values can be provided with comma separated strings
      operationId: StoreModuleFindPetsByStatus
      parameters:
      - name: petStatus
        in: query
        explode: true
        required: true
        description: Status values that need to be considered for filter
        schema:
          type: array
          default: available
          items:
            $ref: '#/components/schemas/PetStatus'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        400:
          description: Invalid status value
  /StoreModule/pet:
    post:
      summary: Add a new pet to the store
      operationId: StoreModuleAddPet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        405:
          description: Invalid input
      x-lz-gencall: PetRepo.CreateAsync(callerInfo, body)
    put:
      summary: Update an existing pet
      operationId: StoreModuleUpdatePet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        400:
          description: Invalid ID supplied
        404:
          description: Pet not found
        405:
          description: Validation exception
      x-lz-gencall: PetRepo.UpdateAsync(callerInfo, body)
  /StoreModule/pet/findByTags:
    get:
      summary: Finds Pets by tags
      description: Muliple tags can be provided with comma separated strings. Use\ \ tag1, tag2, tag3 for testing.
      operationId: StoreModuleFindPetsByTags
      parameters:
      - name: tags
        in: query
        explode: true
        required: true
        description: Tags to filter by
        schema:
          type: array
          items:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        400:
          description: Invalid tag value
  /StoreModule/pet/categories:
    get:
      summary: Get all Pet Categories
      operationId: StoreModuleGetPetCategories
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
      x-lz-gencall: CategoryRepo.ListAsync(callerInfo)
  /StoreModule/pet/tags:
    get:
      summary: Get all Pet Tags
      operationId: StoreModuleGetPetTags
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
      x-lz-gencall: TagRepo.ListAsync(callerInfo)
  /StoreModule/order/inventory:
    get:
      summary: Returns pet inventories by status
      description: Returns a map of status codes to quantities
      operationId: StoreModuleGetInventory
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: integer
                  format: int32
  /StoreModule/order:
    post:
      summary: Place an order for a pet
      operationId: StoreModulePlaceOrder
      requestBody:
        description: order placed for purchasing the pet
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Invalid Order
      x-lz-gencall: OrderRepo.CreateAsync(callerInfo, body)
  /ConsumerModule/preferences:
    get:
      summary: Get user preferences
      operationId: ConsumerModuleGetPreferences
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preferences'
      x-lz-gencall: PreferencesRepo.ReadAsync(callerInfo, callerInfo.LzUserId)
    post:
      summary: Update user preferences
      operationId: ConsumerModuleUpdatePreferences
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Preferences'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preferences'
      x-lz-gencall: PreferencesRepo.UpdateAsync(callerInfo, body)
  /PublicModule/fingerprint/create:
    post:
      operationId: PublicModuleFingerprintCreate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Fingerprint'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fingerprint'
        403:
          description: Invalid permissions
        405:
          description: Invalid data
      x-lz-gencall: FingerprintRepo.CreateAsync(callerInfo, body)
  /PublicModule/pet/listPets:
    get:
      summary: List all pets
      operationId: PublicModuleListPets
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        400:
          description: Invalid status value
      x-lz-gencall: PetRepo.ListAsync(callerInfo)
  /PublicModule/pet/findByStatus:
    get:
      summary: Finds Pets by status
      description: Multiple status values can be provided with comma separated strings
      operationId: PublicModuleFindPetsByStatus
      parameters:
      - name: petStatus
        in: query
        explode: true
        required: true
        description: Status values that need to be considered for filter
        schema:
          type: array
          default: available
          items:
            $ref: '#/components/schemas/PetStatus'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        400:
          description: Invalid status value
  /PublicModule/pet/findByTags:
    get:
      summary: Finds Pets by tags
      description: Muliple tags can be provided with comma separated strings. Use\ \ tag1, tag2, tag3 for testing.
      operationId: PublicModuleFindPetsByTags
      parameters:
      - name: tags
        in: query
        explode: true
        required: true
        description: Tags to filter by
        schema:
          type: array
          items:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        400:
          description: Invalid tag value
  /PublicModule/pet/categories:
    get:
      summary: Get all Pet Categories
      operationId: PublicModuleGetPetCategories
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
      x-lz-gencall: CategoryRepo.ListAsync(callerInfo)
  /PublicModule/pet/tags:
    get:
      summary: Get all Pet Tags
      operationId: PublicModuleGetPetTags
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
      x-lz-gencall: TagRepo.ListAsync(callerInfo)
  '/StoreModule/pet/{petId}':
    get:
      summary: Find pet by ID
      description: Returns a single pet
      operationId: StoreModuleGetPetById
      parameters:
      - name: petId
        in: path
        required: true
        description: ID of pet to return
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        400:
          description: Invalid ID supplied
        404:
          description: Pet not found
      x-lz-gencall: PetRepo.ReadAsync(callerInfo, petId)
    delete:
      summary: Deletes a pet
      operationId: StoreModuleDeletePet
      parameters:
      - name: petId
        in: path
        required: true
        description: Pet id to delete
        schema:
          type: string
      responses:
        200:
          description: Success
        400:
          description: Invalid ID supplied
        404:
          description: Pet not found
      x-lz-gencall: PetRepo.DeleteAsync(callerInfo, petId)
  '/StoreModule/order/{orderId}':
    get:
      summary: Find purchase order by ID
      description: For valid response try integer IDs with value >= 1 and <= 10.\ \ Other values will generated exceptions
      operationId: StoreModuleGetOrderById
      parameters:
      - name: orderId
        in: path
        required: true
        description: ID of pet that needs to be fetched
        schema:
          type: string
          maximum: 10
          minimum: 1
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Invalid ID supplied
        404:
          description: Order not found
      x-lz-gencall: OrderRepo.ReadAsync(callerInfo, orderId)
    delete:
      summary: Delete purchase order by ID
      operationId: StoreModuleDeleteOrder
      parameters:
      - name: orderId
        in: path
        required: true
        description: ID of the order that needs to be deleted
        schema:
          type: string
      responses:
        200:
          description: Success
        400:
          description: Invalid ID supplied
        404:
          description: Order not found
      x-lz-gencall: OrderRepo.DeleteAsync(callerInfo, orderId)
  '/PublicModule/pet/{petId}':
    get:
      summary: Find pet by ID
      description: Returns a single pet
      operationId: PublicModuleGetPetById
      parameters:
      - name: petId
        in: path
        required: true
        description: ID of pet to return
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        400:
          description: Invalid ID supplied
        404:
          description: Pet not found
      x-lz-gencall: PetRepo.ReadAsync(callerInfo, petId)
components:
  schemas:
    Preferences:
      type: object
      properties:
        id:
          type: string
        theme:
          type: string
          enum:
          - light
          - dark
        locale:
          type: string
          enum:
          - en
          - fr
          - es
        petStatus:
          description: pet status in the store
          $ref: '#/components/schemas/PetStatus'
        pets:
          type: array
          items:
            $ref: '#/components/schemas/Pet'
        pet:
          $ref: '#/components/schemas/Pet'
        createUtcTick:
          type: integer
          format: int64
        updateUtcTick:
          type: integer
          format: int64
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      x-lz-genrepo: true
    Tag:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      x-lz-genrepo: true
    Pet:
      type: object
      required:
      - name
      - photoUrls
      properties:
        id:
          type: string
        category:
          type: string
        name:
          type: string
          example: Dog
        photoUrls:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        petStatus:
          description: pet status in the store
          default: available
          $ref: '#/components/schemas/PetStatus'
        createUtcTick:
          type: integer
          format: int64
        updateUtcTick:
          type: integer
          format: int64
      x-lz-genrepo: true
    PetStatus:
      type: string
      description: pet status in the store
      enum:
      - available
      - pending
      - sold
    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
          description: Cognito Identity Id of clerk entering order
        petId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
        shipDate:
          type: string
          format: date-time
        orderStatus:
          title: orderStatus
          description: Order Status
          $ref: '#/components/schemas/OrderStatus'
        complete:
          type: boolean
          default: false
        createUtcTick:
          type: integer
          format: int64
        updateUtcTick:
          type: integer
          format: int64
      x-lz-genrepo: true
    OrderStatus:
      type: string
      description: Order Status
      enum:
      - placed
      - approved
      - delivered
    Subtenant:
      type: object
      x-csharp-type: TenancyConfig
    TenantUser:
      type: object
      properties:
        id:
          type: string
        login:
          type: string
          nullable: true
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        employee:
          type: boolean
        admin:
          type: boolean
        createUtcTick:
          type: integer
          format: int64
        updateUtcTick:
          type: integer
          format: int64
      x-lz-genrepo: true
    TenantUserStatus:
      type: object
      properties:
        isAdmin:
          type: boolean
      x-lz-genrepo: true
    Bada:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      x-lz-genrepo: true
    Fingerprint:
      type: object
      properties:
        id:
          type: string
        fingerprint:
          type: number
        browser:
          type: string
        browserVersion:
          type: string
        os:
          type: string
        osVersion:
          type: string
        resolution:
          type: string
        userAgent:
          type: string
        device:
          type: string
        deviceType:
          type: string
        deviceVendor:
          type: string
        language:
          type: string
        timeZone:
          type: string
        screenPrint:
          type: string
        isMobile:
          type: boolean
        isChrome:
          type: boolean
        isFirefox:
          type: boolean
        isSafari:
          type: boolean
        isIE:
          type: boolean
        createUtcTick:
          type: integer
          format: int64
        updateUtcTick:
          type: integer
          format: int64
...