# Modules/StoreModule/StoreModuleControllerBase.g.cs
// NSWAG code refactored by LazyMagic.
// We use NSWAG to generate a baseclass, partial class and interface. 
// I{ModuleName}Controller.g.cs - defines a partial interface
// {ModuleName}ControllerBase.g.cs - defines the base class
// {ModuleName}Controller.g.cs - defines a partial class that inherits the base class
// To add or override class behavior, create a new partial class file
// {projectName}Controller.cs - overrides methods in the base class
// Dependency Injection system.
// Note: We also generate some helper classes 
// {ModuleName}Authorization.g.cs - Partial class for Authorization system
// {ModuleName}Registration.g.cs - Registers classes with the DI system
//
//----------------------
// <auto-generated>
//     Generated using the NSwag toolchain v14.0.3.0 (NJsonSchema v11.0.0.0 (Newtonsoft.Json v13.0.3.0)) (http://NSwag.org)
// </auto-generated>
//----------------------

namespace StoreModule
{
    using System = global::System;

    [System.CodeDom.Compiler.GeneratedCode("NSwag", "14.0.3.0 (NJsonSchema v11.0.0.0 (Newtonsoft.Json v13.0.3.0))")]

    public abstract class StoreModuleControllerBase : Controller, IStoreModuleController
    {

        /// <summary>
        /// List all pets
        /// </summary>
        /// <returns>successful operation</returns>
        [HttpGet, Route("StoreModule/pet/listPets")]
        public virtual async Task<ActionResult<System.Collections.Generic.ICollection<Pet>>> StoreModuleListPets()
        {
            var callerInfo = await StoreModuleAuthorization.GetCallerInfoAsync(this.Request);
            return await PetRepo.ListAsync(callerInfo);
        }
        /// <summary>
        /// Finds Pets by status
        /// </summary>
        /// <remarks>
        /// Multiple status values can be provided with comma separated strings
        /// </remarks>
        /// <param name="petStatus">Status values that need to be considered for filter</param>
        /// <returns>successful operation</returns>
        [HttpGet, Route("StoreModule/pet/findByStatus")]
        public virtual async Task<ActionResult<System.Collections.Generic.ICollection<Pet>>> StoreModuleFindPetsByStatus([FromQuery] System.Collections.Generic.IEnumerable<PetStatus> petStatus)
        {
            throw new NotImplementedException();
        }
        /// <summary>
        /// Add a new pet to the store
        /// </summary>
        /// <returns>successful operation</returns>
        [HttpPost, Route("StoreModule/pet")]
        public virtual async Task<ActionResult<Pet>> StoreModuleAddPet([FromBody] Pet body)
        {
            var callerInfo = await StoreModuleAuthorization.GetCallerInfoAsync(this.Request);
            return await PetRepo.CreateAsync(callerInfo, body);
        }
        /// <summary>
        /// Update an existing pet
        /// </summary>
        /// <returns>successful operation</returns>
        [HttpPut, Route("StoreModule/pet")]
        public virtual async Task<ActionResult<Pet>> StoreModuleUpdatePet([FromBody] Pet body)
        {
            var callerInfo = await StoreModuleAuthorization.GetCallerInfoAsync(this.Request);
            return await PetRepo.UpdateAsync(callerInfo, body);
        }
        /// <summary>
        /// Finds Pets by tags
        /// </summary>
        /// <remarks>
        /// Muliple tags can be provided with comma separated strings. Use\ \ tag1, tag2, tag3 for testing.
        /// </remarks>
        /// <param name="tags">Tags to filter by</param>
        /// <returns>successful operation</returns>
        [HttpGet, Route("StoreModule/pet/findByTags")]
        public virtual async Task<ActionResult<System.Collections.Generic.ICollection<Pet>>> StoreModuleFindPetsByTags([FromQuery] System.Collections.Generic.IEnumerable<string> tags)
        {
            throw new NotImplementedException();
        }
        /// <summary>
        /// Get all Pet Categories
        /// </summary>
        /// <returns>successful operation</returns>
        [HttpGet, Route("StoreModule/pet/categories")]
        public virtual async Task<ActionResult<System.Collections.Generic.ICollection<Category>>> StoreModuleGetPetCategories()
        {
            var callerInfo = await StoreModuleAuthorization.GetCallerInfoAsync(this.Request);
            return await CategoryRepo.ListAsync(callerInfo);
        }
        /// <summary>
        /// Get all Pet Tags
        /// </summary>
        /// <returns>successful operation</returns>
        [HttpGet, Route("StoreModule/pet/tags")]
        public virtual async Task<ActionResult<System.Collections.Generic.ICollection<Tag>>> StoreModuleGetPetTags()
        {
            var callerInfo = await StoreModuleAuthorization.GetCallerInfoAsync(this.Request);
            return await TagRepo.ListAsync(callerInfo);
        }
        /// <summary>
        /// Returns pet inventories by status
        /// </summary>
        /// <remarks>
        /// Returns a map of status codes to quantities
        /// </remarks>
        /// <returns>successful operation</returns>
        [HttpGet, Route("StoreModule/order/inventory")]
        public virtual async Task<ActionResult<System.Collections.Generic.IDictionary<string, int>>> StoreModuleGetInventory()
        {
            throw new NotImplementedException();
        }
        /// <summary>
        /// Place an order for a pet
        /// </summary>
        /// <param name="body">order placed for purchasing the pet</param>
        /// <returns>successful operation</returns>
        [HttpPost, Route("StoreModule/order")]
        public virtual async Task<ActionResult<Order>> StoreModulePlaceOrder([FromBody] Order body)
        {
            var callerInfo = await StoreModuleAuthorization.GetCallerInfoAsync(this.Request);
            return await OrderRepo.CreateAsync(callerInfo, body);
        }
        /// <summary>
        /// Find pet by ID
        /// </summary>
        /// <remarks>
        /// Returns a single pet
        /// </remarks>
        /// <param name="petId">ID of pet to return</param>
        /// <returns>successful operation</returns>
        [HttpGet, Route("StoreModule/pet/{petId}")]
        public virtual async Task<ActionResult<Pet>> StoreModuleGetPetById(string petId)
        {
            var callerInfo = await StoreModuleAuthorization.GetCallerInfoAsync(this.Request);
            return await PetRepo.ReadAsync(callerInfo, petId);
        }
        /// <summary>
        /// Deletes a pet
        /// </summary>
        /// <param name="petId">Pet id to delete</param>
        /// <returns>Success</returns>
        [HttpDelete, Route("StoreModule/pet/{petId}")]
        public virtual async Task<IActionResult> StoreModuleDeletePet(string petId)
        {
            var callerInfo = await StoreModuleAuthorization.GetCallerInfoAsync(this.Request);
            return await PetRepo.DeleteAsync(callerInfo, petId);
        }
        /// <summary>
        /// Find purchase order by ID
        /// </summary>
        /// <remarks>
        /// For valid response try integer IDs with value &gt;= 1 and &lt;= 10.\ \ Other values will generated exceptions
        /// </remarks>
        /// <param name="orderId">ID of pet that needs to be fetched</param>
        /// <returns>successful operation</returns>
        [HttpGet, Route("StoreModule/order/{orderId}")]
        public virtual async Task<ActionResult<Order>> StoreModuleGetOrderById(string orderId)
        {
            var callerInfo = await StoreModuleAuthorization.GetCallerInfoAsync(this.Request);
            return await OrderRepo.ReadAsync(callerInfo, orderId);
        }
        /// <summary>
        /// Delete purchase order by ID
        /// </summary>
        /// <param name="orderId">ID of the order that needs to be deleted</param>
        /// <returns>Success</returns>
        [HttpDelete, Route("StoreModule/order/{orderId}")]
        public virtual async Task<IActionResult> StoreModuleDeleteOrder(string orderId)
        {
            var callerInfo = await StoreModuleAuthorization.GetCallerInfoAsync(this.Request);
            return await OrderRepo.DeleteAsync(callerInfo, orderId);
        }
		public IStoreModuleAuthorization StoreModuleAuthorization { get; set; }
		public ICategoryRepo CategoryRepo { get; set; }
		public ITagRepo TagRepo { get; set; }
		public IPetRepo PetRepo { get; set; }
		public IOrderRepo OrderRepo { get; set; }
		protected virtual void Init() { }
    }


}