# Schemas/PublicSchema/Models/FingerprintModel.g.cs
//----------------------
// <auto-generated>
//     Generated by LazyMagic.
//     Do not modify, your changes will be overwritten.
//     This is a partial class so you can extend the class in a separate file.
//     Only files with *.g.* extensions will be overwritten.
// </auto-generated>
//----------------------

namespace PublicSchema
{
    using System = global::System;
    

public partial class FingerprintModel : Fingerprint,IRegisterObservables
{
}

}